Added PostViewModel
